pyproject.toml
# Copyright 2025 Henkel Technical Computing, LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

# See PEP-621 for more info

#######################################################################
#                        GENERAL PYTHON CONFIG                        #
#######################################################################
[project]
name = "nexa"
version = "0.1.0"
description = "Material Descriptions for Nuclear Analysis"
keywords = [
    "MCNP",
    "radiation transport",
    "shielding",
    "criticality",
    "activation",
]
license = { file = "LICENSE.txt" }
requires-python = ">=3.9.2"
readme = "README.md"
authors = [{ name = "Charles Henkel", email = "csh@henktech.com" }]
dependencies = [
    "numpy",
    "pandas",
    "scipy",
    "matplotlib",
    "psutil",
    "h5py",
    "ruamel.yaml",
    "voluptuous",
    "yamlize",
]

[build-system]
requires = ["setuptools>=61.2"]
build-backend = "setuptools.build_meta"

[project.scripts]
nexa = "nexa.main:main"
nexa-cli = "nexa.main:main_cli"
getkeff = "nexa.tools.getkeff:main"
getMcnpIso = "nexa.tools.getMcnpIso:main"
getMcnpFlux = "nexa.tools.getMCNPFlux:main"
getOrigenIso = "nexa.tools.getOrigenIso:main"

[tool.ruff]
line-length = 100
